agent_workflow/requirements.txt
# Microsoft Agent Framework (Preview)
agent-framework-azure-ai

# Azure SDKs
azure-identity
azure-search-documents

# OpenAI
openai

# Utilities
python-dotenv
numpy
